Shared Preference, Form validation
Shared preference is added to keep the user always logged in once he has been registered with his mobile number, after signup form submission the user_status is updated to not affected in the database, the user ID is retrieved from the database and stored in the app for future use. Form validation is done so that users don't submit the empty form.